Phase 6: Fix SQLX enum mappings and database schema sync
- Fixed database schema by adding missing turnover_rate column
- Implemented proper SQLX enum handling with string conversion
- Added comprehensive enum parsing for ProductType, ProductStatus, UnitOfMeasure
- Created utils.rs with type conversion utilities
- Added missing analytics engine trait methods
- Reduced compilation errors from 291 to ~12 (96% improvement)

Major accomplishments:
✅ Database schema synchronized with Docker reset
✅ SQLX enum type mappings fixed with PostgreSQL casting
✅ Manual struct mapping for complex type conversions
✅ Analytics engine trait completion

Remaining issues: View insertion errors, missing table columns

+++ crates/master-data/src/supplier/service.rs
@@ -668,8 +668,8 @@ impl SupplierService for MockSupplierService {
         Ok(PaginationResult {
             items,
             total,
-            page: pagination.page,
-            limit: pagination.limit,
+            page: pagination.page() as i64,
+            limit: pagination.limit() as i64,
             total_pages: 1,
         })
     }
